# Faster encoding: ×4 for `byte[]` & ×1,000 for `Stream`