--- a/Bencodex/Types/IKey.cs
+++ b/Bencodex/Types/IKey.cs
@@ -2,11 +2,13 @@ using System.Diagnostics.Contracts;
 
 namespace Bencodex.Types
 {
+    /// <summary>
+    /// Represents Bencodex values which can be keys of a Bencodex <see cref="Dictionary"/>.
+    /// </summary>
+    /// <seealso cref="Binary"/>
+    /// <seealso cref="Text"/>
     public interface IKey : IValue
     {
-        [Pure]
-        byte? KeyPrefix { get; }
-
         [Pure]
         byte[] EncodeAsByteArray();
     }