--- a/Bencodex/Types/Text.cs
+++ b/Bencodex/Types/Text.cs
@@ -159,9 +159,6 @@ namespace Bencodex.Types
             }
         }
 
-        [Pure]
-        byte[] IKey.EncodeAsByteArray() => Encoding.UTF8.GetBytes(Value);
-
         /// <inheritdoc cref="IValue.Inspect(bool)"/>
         public string Inspect(bool loadAll)
         {