This patch fixes a regression made by <https://github.com/planetarium/bencodex.net/pull/49> that `Codec.Encode(IValue, Stream)` requires a stream to be seekable.  Now it accepts any readable streams again.

During this fix, I rewrote the whole encoding logic (so that it does not require streams to be seekable), and accidentally it became drastically faster than before.  Below strategies were applied:

- Instead of a lot of memory allocations for small chunks, it allocates the single long buffer with the exact length of the required bytes for the entire encoded representation.
- When it writes onto a stream, smaller chunks shorter than 4,096 bytes are grouped into a single chunk.
- Unnecessary memcpys were reduced using C#'s `in` keyword for parameters.
- Encoding methods became static in order to avoid dynamic dispatch.

Here's some benchmarks too.  The encoded `Dictionary` weighs about 11 MB.  Benchmarks ran on my MacBook Pro.[^1]

Before optimization (d5cb15bf31b28efe52ee3ce0ec62c6625bf4f7e6):

|           Method |             DataFile |     Mean |    Error |   StdDev |
|----------------- |--------------------- |---------:|---------:|---------:|
|  EncodeIntoBytes | /tmp/(...)e.dat [82] | 32.97 ms | 15.42 ms | 0.845 ms |
| EncodeIntoStream | /tmp/(...)e.dat [82] | 27.24 ms | 19.05 ms | 1.044 ms |

After optimization (a98b62a0d4019f0b2a2cfee7d5c71e8d196c65af):

|           Method |             DataFile |         Mean |        Error |    StdDev |
|----------------- |--------------------- |-------------:|-------------:|----------:|
|  EncodeIntoBytes | /tmp/(...)e.dat [96] |  8,402.07 us | 16,322.05 us | 894.67 us |
| EncodeIntoStream | /tmp/(...)e.dat [96] |     26.45 us |    211.34 us |  11.58 us |

To sum up:

- `Codec.Encode(IValue)` is about 4 times faster than before.
- `Codec.Encdoe(IValue, Stream)` is about 1,000 times faster than before.

[^1]: Intel Core i7 2.2GHz. 16GB 2400MHz DDR4. macOS 11.6. .NET Core SDK 3.1.411.